component decomposition MiniNavigation
1 creating a generic component LinkList
2 added a Linklist element to the BurgerMenu component
